+++ config/assets.hd.yaml
@@ -9,24 +9,24 @@ files:
         vram: 0x800A0000
         assets:
           - [0x0, skip]
-          - [0xD298, blueprintsdef, blueprints, AluBlueprints, AluEntities]
+          - [0xD298, blueprintsdef, hd/blueprints, AluBlueprints, AluEntities]
           - [0xD57C, skip]
-          - [0x11140, binary, vh_0, vh]
-          - [0x14140, binary, vh_1, vh]
-          - [0x16140, binary, vh_2, vh]
-          - [0x18140, binary, vh_3, vh]
+          - [0x11140, binary, hd/vh_0, vh]
+          - [0x14140, binary, hd/vh_1, vh]
+          - [0x16140, binary, hd/vh_2, vh]
+          - [0x18140, binary, hd/vh_3, vh]
           - [0x1A140, binary, sd_eve1, seq, --skip-build]
           - [0x1C441, binary, sd_eve3, seq, --skip-build]
           - [0x1CB87, binary, sd_eve2, seq, --skip-build]
           - [0x1D140, skip]
-          - [0x1D2E8, xamusicconfig, xa_music_config]
-          - [0x1F5F8, sfxconfig, sfx_config]
+          - [0x1D2E8, xamusicconfig, hd/xa_music_config]
+          - [0x1F5F8, sfxconfig, hd/sfx_config]
           - [0x20A20, skip]
-          - [0x22220, cmpgfx, text_item_up, 128, 128, 4]
-          - [0x22854, cmpgfx, load_room_obj, 128, 128, 4]
-          - [0x23604, cmpgfx, save_room_obj, 128, 128, 4]
-          - [0x24908, cmpgfx, load_room_tile, 80, 16, 4]
-          - [0x24B34, cmpgfx, save_room_tile, 80, 16, 4]
+          - [0x22220, cmpgfx, hd/text_item_up, 128, 128, 4]
+          - [0x22854, cmpgfx, hd/load_room_obj, 128, 128, 4]
+          - [0x23604, cmpgfx, hd/save_room_obj, 128, 128, 4]
+          - [0x24908, cmpgfx, hd/load_room_tile, 80, 16, 4]
+          - [0x24B34, cmpgfx, hd/save_room_tile, 80, 16, 4]
           - [0x24D90, headergfx, D_800C4CEC]
           - [0x24E14, headergfx, D_800C4D70]
           - [0x24E98, headergfx, D_800C4DF4]
@@ -41,20 +41,20 @@ files:
           - [0x2539C, skip]
           - [0x253C8, rawgfx, g_GfxEquipIcon, 16, 5120, 4]
           - [0x2F3C8, skip]
-          - [0x2FFB4, spriteset, D_800CFF10]
-          - [0x3100C, spriteset, D_800D0F68]
-          - [0x32C98, spriteset, D_800D2BF4]
-          - [0x32D9C, spriteset, D_800D2CF8]
-          - [0x330A0, spriteset, D_800D2FFC]
-          - [0x33320, spriteset, D_800D327C]
-          - [0x334CC, spriteset, D_800D3428]
-          - [0x3355C, spriteset, D_800D34B8]
-          - [0x337D0, spriteset, D_800D372C]
-          - [0x33C2C, spriteset, D_800D3B88]
-          - [0x356CC, spriteset, D_800D5628]
-          - [0x35750, spriteset, D_800D56AC]
-          - [0x3590C, spriteset, D_800D5868]
-          - [0x368F0, spriteset, D_800D684C]
+          - [0x2FFB4, spriteset, hd/D_800CFF10]
+          - [0x3100C, spriteset, hd/D_800D0F68]
+          - [0x32C98, spriteset, hd/D_800D2BF4]
+          - [0x32D9C, spriteset, hd/D_800D2CF8]
+          - [0x330A0, spriteset, hd/D_800D2FFC]
+          - [0x33320, spriteset, hd/D_800D327C]
+          - [0x334CC, spriteset, hd/D_800D3428]
+          - [0x3355C, spriteset, hd/D_800D34B8]
+          - [0x337D0, spriteset, hd/D_800D372C]
+          - [0x33C2C, spriteset, hd/D_800D3B88]
+          - [0x356CC, spriteset, hd/D_800D5628]
+          - [0x35750, spriteset, hd/D_800D56AC]
+          - [0x3590C, spriteset, hd/D_800D5868]
+          - [0x368F0, spriteset, hd/D_800D684C]
           - [0x36978, palette, D_800D68D4, 16]
           - [0x36998, palette, D_800D68F4, 16]
           - [0x369B8, palette, D_800D6914, 16]
@@ -91,7 +91,7 @@ files:
           - [0x170AC, skip]
           - [0x18688, subweaponsdef, subweapons, RicSubweapons, RicBlueprints]
           - [0x188F4, skip]
-          - [0x18A6C, blueprintsdef, blueprints, RicBlueprints, RicEntities]
+          - [0x18A6C, blueprintsdef, hd/blueprints, RicBlueprints, RicEntities]
           - [0x18C42, skip]
   - target: disks/pspeu/PSP_GAME/USRDIR/res/ps/hdbin/cen.bin
     asset_path: assets/st/cen
