+++ config/assets.pspeu.yaml
@@ -61,16 +61,16 @@ files:
           - [0x84A08, palette, D_800D6ED4, 16]
           - [0x84C08, palette, D_800D70D4, 16]
           - [0x84E08, palette, D_800D72D4, 16]
-          - [0x86408, cmpgfx, text_item_up, 128, 128, 4, 0x86A40]
+          - [0x86408, cmpgfx, psp/text_item_up, 128, 128, 4, 0x86A40]
           - [0x86A40, palette, D_800DB0D4, 16]
           - [0x86A60, palette, D_800DB114, 16]
           - [0x86B40, palette, g_PalEquipIcon, 16]
           - [0x88B40, palette, D_psp_091654C0, 16]
-          - [0x89340, cmpgfx, load_room_obj, 128, 128, 4]
+          - [0x89340, cmpgfx, psp/load_room_obj, 128, 128, 4]
           - [0x8A0F0, palette, D_800DB0F4, 16]
-          - [0x8A110, cmpgfx, save_room_tile, 80, 16, 4]
-          - [0x8A370, cmpgfx, load_room_tile, 80, 16, 4]
-          - [0x8A5A0, cmpgfx, save_room_obj, 128, 128, 4]
+          - [0x8A110, cmpgfx, psp/save_room_tile, 80, 16, 4]
+          - [0x8A370, cmpgfx, psp/load_room_tile, 80, 16, 4]
+          - [0x8A5A0, cmpgfx, psp/save_room_obj, 128, 128, 4]
           - [0x8B8A8, spriteset, D_800D34B8]
           - [0x8BB38, spriteset, D_800D3B88]
           - [0x8D6E0, spriteset, D_800CFF10]
@@ -107,33 +107,33 @@ files:
           - [0xA4540, cmpgfx, D_psp_09180EC0, 128, 128, 4]
           - [0xA56A8, cmpgfx, D_psp_09182028, 128, 128, 4]
           - [0xA67B8, skip]
-          - [0xA6FD8, blueprintsdef, blueprints, AluBlueprints, AluEntities]
+          - [0xA6FD8, blueprintsdef, psp/blueprints, AluBlueprints, AluEntities]
           - [0xA72BC, skip]
           - [0xA8948, palette, g_saveIconPal0, 256]
-          - [0xA8B48, rawgfx, memcard_0, 16, 48, 4, 0xA8948]
-          - [0xA8CC8, rawgfx, memcard_1, 16, 48, 4, 0xA8968]
-          - [0xA8E48, rawgfx, memcard_2, 16, 48, 4, 0xA8988]
-          - [0xA8FC8, rawgfx, memcard_3, 16, 48, 4, 0xA89A8]
-          - [0xA9148, rawgfx, memcard_4, 16, 48, 4, 0xA89C8]
-          - [0xA92C8, rawgfx, memcard_5, 16, 48, 4, 0xA89E8]
-          - [0xA9448, rawgfx, memcard_6, 16, 48, 4, 0xA8A08]
-          - [0xA95C8, rawgfx, memcard_7, 16, 48, 4, 0xA8A28]
-          - [0xA9748, rawgfx, memcard_8, 16, 48, 4, 0xA8A48]
-          - [0xA98C8, rawgfx, memcard_9, 16, 48, 4, 0xA8A68]
-          - [0xA9A48, rawgfx, memcard_10, 16, 48, 4, 0xA8A88]
-          - [0xA9BC8, rawgfx, memcard_11, 16, 48, 4, 0xA8AA8]
-          - [0xA9D48, rawgfx, memcard_12, 16, 48, 4, 0xA8AC8]
-          - [0xA9EC8, rawgfx, memcard_13, 16, 48, 4, 0xA8AE8]
-          - [0xAA048, rawgfx, memcard_14, 16, 48, 4, 0xA8B08]
-          - [0xAA1C8, rawgfx, memcard_15, 16, 48, 4, 0xA8B28]
+          - [0xA8B48, rawgfx, psp/memcard_0, 16, 48, 4, 0xA8948]
+          - [0xA8CC8, rawgfx, psp/memcard_1, 16, 48, 4, 0xA8968]
+          - [0xA8E48, rawgfx, psp/memcard_2, 16, 48, 4, 0xA8988]
+          - [0xA8FC8, rawgfx, psp/memcard_3, 16, 48, 4, 0xA89A8]
+          - [0xA9148, rawgfx, psp/memcard_4, 16, 48, 4, 0xA89C8]
+          - [0xA92C8, rawgfx, psp/memcard_5, 16, 48, 4, 0xA89E8]
+          - [0xA9448, rawgfx, psp/memcard_6, 16, 48, 4, 0xA8A08]
+          - [0xA95C8, rawgfx, psp/memcard_7, 16, 48, 4, 0xA8A28]
+          - [0xA9748, rawgfx, psp/memcard_8, 16, 48, 4, 0xA8A48]
+          - [0xA98C8, rawgfx, psp/memcard_9, 16, 48, 4, 0xA8A68]
+          - [0xA9A48, rawgfx, psp/memcard_10, 16, 48, 4, 0xA8A88]
+          - [0xA9BC8, rawgfx, psp/memcard_11, 16, 48, 4, 0xA8AA8]
+          - [0xA9D48, rawgfx, psp/memcard_12, 16, 48, 4, 0xA8AC8]
+          - [0xA9EC8, rawgfx, psp/memcard_13, 16, 48, 4, 0xA8AE8]
+          - [0xAA048, rawgfx, psp/memcard_14, 16, 48, 4, 0xA8B08]
+          - [0xAA1C8, rawgfx, psp/memcard_15, 16, 48, 4, 0xA8B28]
           - [0xAA348, skip]
           - [0xAD4E0, cmpgfx, D_psp_09189E60, 128, 128, 4, 0x86A40]
           - [0xADD10, cmpgfx, D_psp_0918A690, 128, 128, 4, 0x86A40]
           - [0xAE228, cmpgfx, D_psp_0918ABA8, 128, 128, 4, 0x86A40]
           - [0xAE820, cmpgfx, D_psp_0918B1A0, 128, 128, 4, 0x86A40]
           - [0xAEEB0, skip]
-          - [0xAF3E0, xamusicconfig, xa_music_config]
-          - [0xB16F0, sfxconfig, sfx_config]
+          - [0xAF3E0, xamusicconfig, psp/xa_music_config]
+          - [0xB16F0, sfxconfig, psp/sfx_config]
           - [0xB2C40, skip]
   - target: disks/pspeu/PSP_GAME/USRDIR/res/ps/PSPBIN/ric.bin
     asset_path: assets/ric
@@ -153,7 +153,7 @@ files:
           - [0x2D1B0, skip]
           - [0x2D320, subweaponsdef, subweapons, RicSubweapons, RicBlueprints]
           - [0x2D590, skip]
-          - [0x2D718, blueprintsdef, blueprints, RicBlueprints, RicEntities]
+          - [0x2D718, blueprintsdef, psp/blueprints, RicBlueprints, RicEntities]
           - [0x2D8F0, skip]
   - target: disks/pspeu/PSP_GAME/USRDIR/res/ps/PSPBIN/maria.bin
     asset_path: assets/maria
