+++ config/assets.us.yaml
@@ -9,24 +9,24 @@ files:
         vram: 0x800A0000
         assets:
           - [0x0, skip]
-          - [0xD1D4, blueprintsdef, blueprints, AluBlueprints, AluEntities]
+          - [0xD1D4, blueprintsdef, us/blueprints, AluBlueprints, AluEntities]
           - [0xD4B8, skip]
-          - [0x1107C, binary, vh_0, vh]
-          - [0x1407C, binary, vh_1, vh]
-          - [0x1607C, binary, vh_2, vh]
-          - [0x1807C, binary, vh_3, vh]
+          - [0x1107C, binary, us/vh_0, vh]
+          - [0x1407C, binary, us/vh_1, vh]
+          - [0x1607C, binary, us/vh_2, vh]
+          - [0x1807C, binary, us/vh_3, vh]
           - [0x1A07C, binary, sd_eve1, seq, --skip-build]
           - [0x1C37D, binary, sd_eve3, seq, --skip-build]
           - [0x1CAC3, binary, sd_eve2, seq, --skip-build]
           - [0x1D07C, skip]
-          - [0x1D224, xamusicconfig, xa_music_config]
-          - [0x1F554, sfxconfig, sfx_config]
+          - [0x1D224, xamusicconfig, us/xa_music_config]
+          - [0x1F554, sfxconfig, us/sfx_config]
           - [0x2097B, skip]
-          - [0x2217C, cmpgfx, text_item_up, 128, 128, 4]
-          - [0x227B0, cmpgfx, load_room_obj, 128, 128, 4]
-          - [0x23560, cmpgfx, save_room_obj, 128, 128, 4]
-          - [0x24864, cmpgfx, load_room_tile, 80, 16, 4]
-          - [0x24A90, cmpgfx, save_room_tile, 80, 16, 4]
+          - [0x2217C, cmpgfx, us/text_item_up, 128, 128, 4]
+          - [0x227B0, cmpgfx, us/load_room_obj, 128, 128, 4]
+          - [0x23560, cmpgfx, us/save_room_obj, 128, 128, 4]
+          - [0x24864, cmpgfx, us/load_room_tile, 80, 16, 4]
+          - [0x24A90, cmpgfx, us/save_room_tile, 80, 16, 4]
           - [0x24CEC, headergfx, D_800C4CEC]
           - [0x24D70, headergfx, D_800C4D70]
           - [0x24DF4, headergfx, D_800C4DF4]
@@ -41,20 +41,20 @@ files:
           - [0x252F8, skip]
           - [0x25324, rawgfx, g_GfxEquipIcon, 16, 5120, 4]
           - [0x2F324, skip]
-          - [0x2FF10, spriteset, D_800CFF10]
-          - [0x30F68, spriteset, D_800D0F68]
-          - [0x32BF4, spriteset, D_800D2BF4]
-          - [0x32CF8, spriteset, D_800D2CF8]
-          - [0x32FFC, spriteset, D_800D2FFC]
-          - [0x3327C, spriteset, D_800D327C]
-          - [0x33428, spriteset, D_800D3428]
-          - [0x334B8, spriteset, D_800D34B8]
-          - [0x3372C, spriteset, D_800D372C]
-          - [0x33B88, spriteset, D_800D3B88]
-          - [0x35628, spriteset, D_800D5628]
-          - [0x356AC, spriteset, D_800D56AC]
-          - [0x35868, spriteset, D_800D5868]
-          - [0x3684C, spriteset, D_800D684C]
+          - [0x2FF10, spriteset, us/D_800CFF10]
+          - [0x30F68, spriteset, us/D_800D0F68]
+          - [0x32BF4, spriteset, us/D_800D2BF4]
+          - [0x32CF8, spriteset, us/D_800D2CF8]
+          - [0x32FFC, spriteset, us/D_800D2FFC]
+          - [0x3327C, spriteset, us/D_800D327C]
+          - [0x33428, spriteset, us/D_800D3428]
+          - [0x334B8, spriteset, us/D_800D34B8]
+          - [0x3372C, spriteset, us/D_800D372C]
+          - [0x33B88, spriteset, us/D_800D3B88]
+          - [0x35628, spriteset, us/D_800D5628]
+          - [0x356AC, spriteset, us/D_800D56AC]
+          - [0x35868, spriteset, us/D_800D5868]
+          - [0x3684C, spriteset, us/D_800D684C]
           - [0x368D4, palette, D_800D68D4, 16]
           - [0x368F4, palette, D_800D68F4, 16]
           - [0x36914, palette, D_800D6914, 16]
@@ -91,7 +91,7 @@ files:
           - [0x170AC, skip]
           - [0x18688, subweaponsdef, subweapons, RicSubweapons, RicBlueprints]
           - [0x188F4, skip]
-          - [0x18A6C, blueprintsdef, blueprints, RicBlueprints, RicEntities]
+          - [0x18A6C, blueprintsdef, us/blueprints, RicBlueprints, RicEntities]
           - [0x18C42, skip]
   - target: disks/us/ST/CEN/CEN.BIN
     asset_path: assets/st/cen
