+++ config/splat.hd.dra.yaml
@@ -40,38 +40,38 @@ segments:
       - [0x160, .data, game_handlers]
       - [0x248, .data, 47BB8]
       - [0x510, .data, save_mgr_pre]
-      - [0x518, palette, memcard_0]
-      - [0x538, palette, memcard_1]
-      - [0x558, palette, memcard_2]
-      - [0x578, palette, memcard_3]
-      - [0x598, palette, memcard_4]
-      - [0x5B8, palette, memcard_5]
-      - [0x5D8, palette, memcard_6]
-      - [0x5F8, palette, memcard_7]
-      - [0x618, palette, memcard_8]
-      - [0x638, palette, memcard_9]
-      - [0x658, palette, memcard_10]
-      - [0x678, palette, memcard_11]
-      - [0x698, palette, memcard_12]
-      - [0x6B8, palette, memcard_13]
-      - [0x6D8, palette, memcard_14]
-      - [0x6F8, palette, memcard_15]
-      - [0x718, ci4, memcard_0, 16, 48]
-      - [0x898, ci4, memcard_1, 16, 48]
-      - [0xA18, ci4, memcard_2, 16, 48]
-      - [0xB98, ci4, memcard_3, 16, 48]
-      - [0xD18, ci4, memcard_4, 16, 48]
-      - [0xE98, ci4, memcard_5, 16, 48]
-      - [0x1018, ci4, memcard_6, 16, 48]
-      - [0x1198, ci4, memcard_7, 16, 48]
-      - [0x1318, ci4, memcard_8, 16, 48]
-      - [0x1498, ci4, memcard_9, 16, 48]
-      - [0x1618, ci4, memcard_10, 16, 48]
-      - [0x1798, ci4, memcard_11, 16, 48]
-      - [0x1918, ci4, memcard_12, 16, 48]
-      - [0x1A98, ci4, memcard_13, 16, 48]
-      - [0x1C18, ci4, memcard_14, 16, 48]
-      - [0x1D98, ci4, memcard_15, 16, 48]
+      - [0x518, palette, hd/memcard_0]
+      - [0x538, palette, hd/memcard_1]
+      - [0x558, palette, hd/memcard_2]
+      - [0x578, palette, hd/memcard_3]
+      - [0x598, palette, hd/memcard_4]
+      - [0x5B8, palette, hd/memcard_5]
+      - [0x5D8, palette, hd/memcard_6]
+      - [0x5F8, palette, hd/memcard_7]
+      - [0x618, palette, hd/memcard_8]
+      - [0x638, palette, hd/memcard_9]
+      - [0x658, palette, hd/memcard_10]
+      - [0x678, palette, hd/memcard_11]
+      - [0x698, palette, hd/memcard_12]
+      - [0x6B8, palette, hd/memcard_13]
+      - [0x6D8, palette, hd/memcard_14]
+      - [0x6F8, palette, hd/memcard_15]
+      - [0x718, ci4, hd/memcard_0, 16, 48]
+      - [0x898, ci4, hd/memcard_1, 16, 48]
+      - [0xA18, ci4, hd/memcard_2, 16, 48]
+      - [0xB98, ci4, hd/memcard_3, 16, 48]
+      - [0xD18, ci4, hd/memcard_4, 16, 48]
+      - [0xE98, ci4, hd/memcard_5, 16, 48]
+      - [0x1018, ci4, hd/memcard_6, 16, 48]
+      - [0x1198, ci4, hd/memcard_7, 16, 48]
+      - [0x1318, ci4, hd/memcard_8, 16, 48]
+      - [0x1498, ci4, hd/memcard_9, 16, 48]
+      - [0x1618, ci4, hd/memcard_10, 16, 48]
+      - [0x1798, ci4, hd/memcard_11, 16, 48]
+      - [0x1918, ci4, hd/memcard_12, 16, 48]
+      - [0x1A98, ci4, hd/memcard_13, 16, 48]
+      - [0x1C18, ci4, hd/memcard_14, 16, 48]
+      - [0x1D98, ci4, hd/memcard_15, 16, 48]
       - [0x1F18, .data, save_mgr]
       - [0x2178, .data, 4B758] # D_800A21B8
       - [0x23F8, .data, 4DA70] # D_800A2438
