+++ src/dra/d_2217C.c
@@ -1,18 +1,18 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
-#include <game.h>
+#include <common.h>
 
 u8 text_item_up[] = {
-#include "gen/text_item_up.h"
+#include GEN_VERSION(text_item_up.h)
 };
 u8 load_room_obj[] = {
-#include "gen/load_room_obj.h"
+#include GEN_VERSION(load_room_obj.h)
 };
 u8 save_room_obj[] = {
-#include "gen/save_room_obj.h"
+#include GEN_VERSION(save_room_obj.h)
 };
 u8 load_room_tile[] = {
-#include "gen/load_room_tile.h"
+#include GEN_VERSION(load_room_tile.h)
 };
 u8 save_room_tile[] = {
-#include "gen/save_room_tile.h"
+#include GEN_VERSION(save_room_tile.h)
 };
