+++ src/ric/pl_blueprints.c
@@ -560,7 +560,7 @@ Entity* RicCreateEntFactoryFromEntity(
 }
 
 static FactoryBlueprint blueprints[] = {
-#include "gen/blueprints.h"
+#include GEN_VERSION(blueprints.h)
 };
 STATIC_ASSERT(LEN(blueprints) == NUM_BLUEPRINTS, "bp array wrong size");
